Refactor out base class
