+++ Assets/MRTK/Providers/WindowsMixedReality/Shared/BaseWindowsMixedRealityEyeGazeDataProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 637f36da999361040843f5ff8d27dce5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
